Phase_V/services/recurring-service/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.27.0
python-dateutil>=2.9.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.0
prometheus-fastapi-instrumentator>=6.1.0
python-dotenv>=1.0.0
